+++ src/apis/app/app.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
 import { APP_GUARD } from '@nestjs/core';
-import { AppController } from './app.controller';
-import { PrismaModule } from '../prisma/prisma.module';
-import { BookmarkModule } from '../bookmark/bookmark.module';
-import { TrashModule } from '../trash/trash.module';
-import { TagModule } from '../tag/tag.module';
-import { AuthModule } from '../auth/auth.module';
-import { UserModule } from '../user/user.module';
+import { PrismaModule } from '@/prisma';
+import { AppController } from '@/apis/app';
+import { BookmarkModule } from '@/apis/bookmark';
+import { TrashModule } from '@/apis/trash';
+import { TagModule } from '@/apis/tag';
+import { AuthModule } from '@/apis/auth';
+import { UserModule } from '@/apis/user';
 
 import { JwtAuthGuard } from '@/helpers/guard';
