+++ src/apis/app/index.ts
@@ -1,0 +1,2 @@
+export { AppModule } from './app.module';
+export { AppController } from './app.controller';
