+++ src/apis/auth/auth.module.ts
@@ -2,9 +2,8 @@ import { Module } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
-import { UserService } from '../user/user.service';
-import { AuthService } from './auth.service';
-import { AuthController } from './auth.controller';
+import { UserService } from '@/apis/user';
+import { AuthService, AuthController } from '@/apis/auth';
 import {
   GoogleStrategy,
   KakaoStrategy,
