+++ src/apis/auth/index.ts
@@ -1,0 +1,3 @@
+export { AuthModule } from './auth.module';
+export { AuthController } from './auth.controller';
+export { AuthService } from './auth.service';
