+++ src/apis/bookmark/bookmark.module.ts
@@ -1,7 +1,6 @@
 import { Module } from '@nestjs/common';
-import { BookmarkController } from './bookmark.controller';
-import { BookmarkService } from './bookmark.service';
-import { TagModule } from '../tag/tag.module';
+import { BookmarkController, BookmarkService } from '@/apis/bookmark';
+import { TagModule } from '@/apis/tag';
 
 @Module({
   imports: [TagModule],
