+++ src/apis/bookmark/index.ts
@@ -1,0 +1,3 @@
+export { BookmarkModule } from './bookmark.module';
+export { BookmarkController } from './bookmark.controller';
+export { BookmarkService } from './bookmark.service';
