+++ src/apis/tag/index.ts
@@ -1,0 +1,3 @@
+export { TagModule } from './tag.module';
+export { TagController } from './tag.controller';
+export { TagService } from './tag.service';
