+++ src/apis/trash/index.ts
@@ -1,0 +1,3 @@
+export { TrashModule } from './trash.module';
+export { TrashController } from './trash.controller';
+export { TrashService } from './trash.service';
