+++ src/apis/user/index.ts
@@ -1,0 +1,3 @@
+export { UserModule } from './user.module';
+export { UserController } from './user.controller';
+export { UserService } from './user.service';
