+++ src/apis/user/user.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
-import { UserService } from './user.service';
-import { UserController } from './user.controller';
+import { UserService, UserController } from '@/apis/user';
 
 @Module({
   providers: [UserService],
